move token transfer transaction to testutils
